+++ scilab/modules/linear_algebra/sci_gateway/cpp/sci_balanc.cpp
@@ -1,5 +1,5 @@
 /*
-* Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
+* Scilab ( https://www.scilab.org/ ) - This file is part of Scilab
 * Copyright (C) 2009 - DIGITEO - Bernard HUGUENEY
 * Copyright (C) 2011 - DIGITEO - Cedric DELAMARRE
 *
@@ -44,7 +44,7 @@ types::Function::ReturnValue sci_balanc(types::typed_list &in, int _iRetCount, t
 
     if (_iRetCount != 2 * in.size())
     {
-        Scierror(78, _("%s: Wrong number of output argument(s): %d expected.\n"), "balanc", 2 * in.size());
+        Scierror(78, _("%s: Wrong number of output argument(s): %d expected.\n"), "balanc", 2 * (int) in.size());
         return types::Function::Error;
     }
